finalização do axios
