+++ src/router/index.js
@@ -2,7 +2,6 @@ import { createRouter, createWebHistory } from "vue-router";
 import NavBar from "../views/NavBarView.vue";
 import HomePage from "../views/HomePageView.vue";
 import RegistrationUser from "../views/RegistrationUserView.vue";
-import AddressUser from "@/components/AddressUser.vue";
 import BoxSwipe from "@/components/BoxSwipe.vue";
 import Login from "@/views/LoginView.vue";
 
@@ -23,11 +22,6 @@ const router = createRouter({
       path: "/registration",
       name: "registration",
       component: RegistrationUser,
-    },
-    {
-      path: "/address",
-      name: "address",
-      component: AddressUser,
     },
   ],
 });
